code migrated from ProdectFetch to products page
