+++ src/app/products/page.jsx
@@ -1,9 +1,46 @@
-import React from 'react'
+'use client'
 
-const page = () => {
+import { useEffect, useState } from "react"
+
+
+ const ProductFetch = () => {
+
+    const [product, setProduct] = useState([null]);
+
+    
+
+    useEffect(() => {
+      async function fetchData() {
+        const res = await fetch('https://dummyjson.com/products');
+        const data = await res.json();
+        
+        setProduct(data.products);
+      }
+  
+      fetchData();
+    }, []);
+    
+    console.log("2nd log", product)
+
   return (
-    <div>page</div>
+   <>
+   <h1>halo</h1>
+
+    <div>
+      <h1>List of Items</h1>
+      <ul>
+        {product.map((item) => (
+          <li key={item?.id}>
+            <h2>{item?.title}</h2>
+            <p>{item?.description}</p>
+          </li>
+        ))}
+      </ul>
+    </div>
+   </>
+    
   )
 }
 
-export default page
+
+export default ProductFetch
